This is the initial commit for this project.
This is the inital commit to my Summer 2013 ResNet project. This is an
online ticket system in which residents of Humboldt State Housing may
visit and after entering required personal information, may describe
their problem, and then submit the ticket to the Housing department. Any
attendant at the front desk will receive a notification  and after
reviewing the students request  can submit it to the  maintenance staff.
This project interacts with FileMakerPro as an external entity.